refactor: 테스트 코드 도메인 분리
