+++ __tests__/WinningLottoTest.test.js
@@ -1,4 +1,5 @@
 import { ERROR_MESSAGE } from "../src/constants";
+import Lotto from "../src/Model/Lotto";
 import WinningLotto from "../src/Model/WinningLotto";
 
 describe("당첨 로또 테스트", () => {
@@ -72,5 +73,29 @@ describe("당첨 로또 테스트", () => {
     expect(() => new WinningLotto(winningNUmbers, wrongBonusNumber)).toThrow(
       ERROR_MESSAGE.PREFIX,
     );
+  });
+
+  test("[기능] 구매한 로또 하나와 당첨로또를 비교해서 일치하는 번호의 개수를 반환해야한다", () => {
+    // given
+    const lotto = new Lotto([1, 2, 3, 4, 8, 9]);
+    const winningLotto = new WinningLotto([1, 2, 3, 4, 5, 6], 7);
+
+    // when
+    const matchCount = winningLotto.getMatchCount(lotto);
+
+    // then
+    expect(matchCount).toEqual(4);
+  });
+
+  test("[기능] 구매한 로또 번호에 보너스 번호가 포함되어 있는지 확인해야 한다", () => {
+    // given
+    const lotto = new Lotto([1, 2, 3, 4, 8, 9]);
+    const winningLotto = new WinningLotto([1, 2, 3, 4, 5, 6], 9);
+
+    // when
+    const hasBonus = winningLotto.hasBonus(lotto);
+
+    // then
+    expect(hasBonus).toEqual(true);
   });
 });
